pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = "Example LangGraph project for deployment to LangGraph Cloud"
authors = [
    "Ajinath Ghodake"
]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langchain-community = "^0.2.12"
langchain-ollama = "^0.1.1"
langchain-groq = "^0.1.9"
python-dotenv = "^1.0.1"
tavily-python = "^0.4.0"
langgraph = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-experimental = "^0.0.64"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"